requirements.txt
geopandas==0.12.2
geopy==2.3.0
gevent==22.10.2
locustio==0.999
matplotlib==3.7.1
mimesis==7.0.0
pandas==1.5.3
plotly_express==0.4.1
pymongo==4.3.3
requests==2.28.2
tqdm==4.65.0